segments: Add missing API for "dot" control on big 7-segments
There are 4 side LEDs around the 4 7-segments on 1.2" packages of HT16k33
(https://www.adafruit.com/product/1270). This commit adds an API to control
them (set and get).

There already was setter and getter to control "ampm" dot (the one at the
top-right). This API is still working but now use the new added functions.