Fixing matrix.fill color flipping
Swapping the _set_buffer() parameters so that when the matrix is filled it shows the proper color. This fixes part of issue #102. Tested with a QT Py RP2040 running CP8 and an 8x8 BiColor matrix

I tried working on shifting to get the color to remain as expected but that's a little over my head at the moment. Also changing two ValueError() strings to f strings per pylint